add check if parcel is enabled; S2CProposalDeliveredAnswerPacket status parameter change from short to byte; add proposal item check & response if already couple or item not existing

+++ src/main/java/com/jftse/emulator/server/game/core/packet/packets/messaging/S2CProposalDeliveredAnswerPacket.java
@@ -8,7 +8,7 @@ import lombok.Setter;
 @Getter
 @Setter
 public class S2CProposalDeliveredAnswerPacket extends Packet {
-    public S2CProposalDeliveredAnswerPacket(short status) {
+    public S2CProposalDeliveredAnswerPacket(byte status) {
         super(PacketID.S2CProposalDeliveredAnswer);
 
         this.write(status);
